switch for keeping temp dirs

+++ src/toolkit/base/case.cpp
@@ -26,7 +26,8 @@ using namespace boost::filesystem;
 namespace insight
 {
   
-TemporaryCaseDir::TemporaryCaseDir()
+TemporaryCaseDir::TemporaryCaseDir(bool keep)
+: keep_(keep)
 {
   dir = unique_path();
   create_directories(dir);
@@ -34,7 +35,8 @@ TemporaryCaseDir::TemporaryCaseDir()
 
 TemporaryCaseDir::~TemporaryCaseDir()
 {
-  remove_all(dir);
+  if (!keep_)
+    remove_all(dir);
 }
